chore(warehouse-sources): move the SCD2 writer into core/delta
Phase 3 (PR 2) of the warehouse-sources untangling, after the maintenance
split (#75978): the CDC-only SCD2 write path leaves DeltaTableHelper.

- core/delta/scd2.py: Scd2DeltaWriter, a stateless wrapper over the table
  helper. The validity-interval column names are injected by the caller
  (the CDC load processor passes the columns cdc/batcher.py stamps), so
  the writer itself carries nothing CDC-specific and can be reused when
  CDC becomes a source-oriented ingress consumed by the scheduled sync.
- core/delta/evolution.py: schema evolution (evolve_delta_schema) as a
  free function shared by the core write path and the SCD2 writer; kept
  out of delta/ops.py so the dlt import stays off the maintenance-only
  import path.
- delta_merge_spill_kwargs moves to delta/ops.py; the pure-arrow helpers
  realign_decimal_buffers and first_per_pk_table move to core/arrow_utils
  (repartition.py stops importing them from the helper), all dropping
  their now-misleading leading underscores.

No behavior changes; DeltaTableHelper shrinks by ~240 lines and its only
remaining write concern is write_to_deltalake, which PR 3 extracts.
