+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/core/delta/ops.py
@@ -1,9 +1,30 @@
 import asyncio
 from collections.abc import Callable
 
+from django.conf import settings
+
 import deltalake
 import deltalake.exceptions
 from structlog.types import FilteringBoundLogger
+
+
+def delta_merge_spill_kwargs() -> dict[str, int]:
+    """delta-rs `merge` kwargs that let DataFusion spill to disk instead of OOMing on large merges.
+
+    A merge decompresses the target partition into an Arrow working set that can exceed the pod's
+    memory limit and take down every co-tenant activity. When the byte budgets are configured (and the
+    worker mounts a scratch disk at its TMPDIR), delta-rs bounds DataFusion's memory pool: bytes past
+    `max_spill_size` spill to disk, capped at `max_temp_directory_size`. Unset → omit the kwargs so
+    DataFusion keeps its unbounded default (today's behavior), which also keeps this compatible with
+    deltalake versions predating the parameters.
+    """
+    kwargs: dict[str, int] = {}
+    if settings.DATA_WAREHOUSE_DELTA_MERGE_MAX_SPILL_SIZE_BYTES is not None:
+        kwargs["max_spill_size"] = settings.DATA_WAREHOUSE_DELTA_MERGE_MAX_SPILL_SIZE_BYTES
+    if settings.DATA_WAREHOUSE_DELTA_MERGE_MAX_TEMP_DIRECTORY_SIZE_BYTES is not None:
+        kwargs["max_temp_directory_size"] = settings.DATA_WAREHOUSE_DELTA_MERGE_MAX_TEMP_DIRECTORY_SIZE_BYTES
+    return kwargs
+
 
 # Delta's conflict checker raises CommitFailedError the moment a concurrent commit invalidates what
 # a committing operation read — a merge predicate, or optimize.compact's file-rewrite plan — unlike a
